diffusion/Latent Diffusion/Conditioned/CelebA/requirements.txt
torch
torchvision
transformers
gradio
spacy
datasets
Pillow